+++ src/core/server/commands/item.ts
@@ -13,9 +13,7 @@ const pistolItem: Item = {
     uuid: `some_hash_thing_ground`,
     description: `Debug: Should be able to go to toolbar. Cannot go to equipment. Is a weapon.`,
     icon: 'pistol',
-    slot: 0,
     quantity: 1,
-    weight: 2,
     behavior: ItemType.CAN_DROP | ItemType.CAN_TRADE | ItemType.IS_TOOLBAR | ItemType.IS_WEAPON,
     data: {
         hash: 0x1b06d571
@@ -27,9 +25,7 @@ const equipmentItem: Item = {
     uuid: `some_hash_thing_ground`,
     description: `Debug: Can go into the bag section. Cannot go into toolbar.`,
     icon: 'backpack',
-    slot: 1,
     quantity: 1,
-    weight: 2,
     equipment: EquipmentType.BAG,
     behavior: ItemType.CAN_DROP | ItemType.CAN_TRADE | ItemType.IS_EQUIPMENT,
     data: {
@@ -42,9 +38,7 @@ const boxItem: Item = {
     uuid: `some_hash_thing_ground`,
     description: `Debug: Inventory only.`,
     icon: 'crate',
-    slot: 2,
     quantity: 1,
-    weight: 5,
     behavior: ItemType.CAN_DROP | ItemType.CAN_TRADE,
     data: {
         drawable: 0
@@ -56,9 +50,7 @@ const smgItem: Item = {
     uuid: `some_hash_thing_ground`,
     description: `Debug: Should be able to go to toolbar. Cannot go to equipment. Is a weapon.`,
     icon: 'smg',
-    slot: 4,
     quantity: 1,
-    weight: 2,
     behavior: ItemType.CAN_DROP | ItemType.CAN_TRADE | ItemType.IS_TOOLBAR | ItemType.IS_WEAPON,
     data: {
         hash: 0x13532244
@@ -70,9 +62,7 @@ const burgerItem: Item = {
     uuid: `food_1`,
     description: `Debug: Should be able to call an event with this`,
     icon: 'burger',
-    slot: 5,
     quantity: 5,
-    weight: 1,
     behavior: ItemType.CAN_DROP | ItemType.CAN_TRADE | ItemType.CONSUMABLE,
     data: {
         event: 'effect:Heal',
@@ -85,28 +75,23 @@ ChatController.addCommand('dummyitem', '/dummyitem - Get some dummy debug items'
 ChatController.addCommand('getitem', '/getitem [item_name] - Get item by name', Permissions.Admin, handleGetItem);
 
 function handleCommand(player: alt.Player): void {
-    let itemClone = deepCloneObject<Item>(pistolItem);
-    let slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
-    playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
-
-    itemClone = deepCloneObject<Item>(equipmentItem);
-    slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
-    playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
-
-    itemClone = deepCloneObject<Item>(boxItem);
-    slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
-    playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
-
-    itemClone = deepCloneObject<Item>(smgItem);
-    slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
-    playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
-
-    itemClone = deepCloneObject<Item>(burgerItem);
-    slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
-    playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
-
-    playerFuncs.save.field(player, 'inventory', player.data.inventory);
-    playerFuncs.sync.inventory(player);
+    // let itemClone = deepCloneObject<Item>(pistolItem);
+    // let slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
+    // playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
+    // itemClone = deepCloneObject<Item>(equipmentItem);
+    // slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
+    // playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
+    // itemClone = deepCloneObject<Item>(boxItem);
+    // slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
+    // playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
+    // itemClone = deepCloneObject<Item>(smgItem);
+    // slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
+    // playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
+    // itemClone = deepCloneObject<Item>(burgerItem);
+    // slotInfo = playerFuncs.inventory.getFreeInventorySlot(player);
+    // playerFuncs.inventory.inventoryAdd(player, itemClone, slotInfo.slot, slotInfo.tab);
+    // playerFuncs.save.field(player, 'inventory', player.data.inventory);
+    // playerFuncs.sync.inventory(player);
 }
 
 function handleGetItem(player: alt.Player, name: string) {