+++ src/core/server/interface/CategoryData.ts
@@ -1,0 +1,8 @@
+export interface CategoryData {
+    abbrv?: string;
+    name: string;
+    emptyCheck?: Function;
+    getItem?: Function;
+    removeItem?: Function;
+    addItem?: Function;
+}