+++ src/core/server/zTests/inventory.ts
@@ -20,7 +20,6 @@ const smgItem: Item = {
     icon: 'gun',
     slot: 4,
     quantity: 1,
-    weight: 2,
     behavior: ItemType.CAN_DROP | ItemType.CAN_TRADE | ItemType.IS_TOOLBAR | ItemType.IS_WEAPON,
     data: {
         test: 'test',